requirements.txt
requests==2.31.0
pandas==2.1.4
google-generativeai==0.3.1
ta==0.10.2
nltk==3.8.1
python-telegram-bot==20.7
numpy==1.26.2
aiohttp==3.9.1
python-dotenv==1.0.0